Solved the problem 1.3 Urlify
Problem Statement:
Write a method to replace all spaces in a string with '%20'. You may assume that the string has sufficient space at the end to hold the additional characters, and that you are given the "true" length of the string. (Note: If implementing in Java, please use a character array so that you can perform this operation in place.)

Approach:
replacing each iteration of " " with "%20" using replace method.

Time Complexity:
O(n) (Note: 'n' represents the number of blank spaces)

File Added: +1
03_urlify.py

File Modified: 0