Solved the problem 1.4 Palindrome Permutation
Problem Statement:
Given a string, write a function to check if it is a permutation of a palindrome. A palindrome is a word or phrase that is the same forwards and backwards. A permutation is a rearrangement of letters. The palindrome does not need to be limited to just dictionary words.

Data Structure Used:
hash map/dict

Time Complexity:
O(n) (Note: 'n' represents the length of the string)

File Added: +1
04_palindrome_permutation.py

File Modified: 0